pyproject.toml
[project]
name = "cad-3dto2d"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "neka-nat", email = "nekanat.stock@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "build123d==0.8.0",
    "cadquery-ocp>=7.7.0a0,<7.8",
    "pydantic>=2.12.5",
    "pyyaml>=6.0",
    "rlpycairo>=0.3.0",
    "svglib>=1.5.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
